Add configure implementation
