+++ internal/datasources/app_endpoints.go
@@ -1,0 +1,42 @@
+package datasources
+
+import (
+	"context"
+	"fmt"
+	providerschema "github.com/couchbasecloud/terraform-provider-couchbase-capella/internal/schema"
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
+)
+
+// NewAppEndpoint is used in (p *capellaProvider) DataSources for building the provider.
+func NewAppEndpoint() datasource.DataSource {
+	return &AppEndpoint{}
+}
+
+// Metadata returns the App Service CIDRs data source type name.
+func (a *AppEndpoint) Metadata(
+	_ context.Context, req datasource.MetadataRequest, resp *datasource.MetadataResponse,
+) {
+	resp.TypeName = req.ProviderTypeName + "_app_services_cidr"
+}
+
+func (a *AppEndpoint) Configure(ctx context.Context, req datasource.ConfigureRequest, resp *datasource.ConfigureResponse) {
+	if req.ProviderData == nil {
+		return
+	}
+
+	data, ok := req.ProviderData.(*providerschema.Data)
+	if !ok {
+		resp.Diagnostics.AddError(
+			"Unexpected Data Source Configure Type",
+			fmt.Sprintf("Expected *ProviderSourceData, got: %T. Please report this issue to the provider developers.", req.ProviderData),
+		)
+
+		return
+	}
+
+	a.Data = data
+}
+
+func (a *AppEndpoint) Read(ctx context.Context, request datasource.ReadRequest, response *datasource.ReadResponse) {
+	//TODO implement me
+}
